Change Read() size and return value to a size_t for clarity
BUG=100641



TEST=N/A
Review URL: http://codereview.chromium.org/8352019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106533 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/filters/ffmpeg_demuxer.h
@@ -148,8 +148,8 @@ class MEDIA_EXPORT FFmpegDemuxer : public Demuxer, public FFmpegURLProtocol {
   virtual void SetPreload(Preload preload) OVERRIDE;
   virtual base::TimeDelta GetStartTime() const OVERRIDE;
 
-  // FFmpegProtocol implementation.
-  virtual int Read(int size, uint8* data) OVERRIDE;
+  // FFmpegURLProtocol implementation.
+  virtual size_t Read(size_t size, uint8* data) OVERRIDE;
   virtual bool GetPosition(int64* position_out) OVERRIDE;
   virtual bool SetPosition(int64 position) OVERRIDE;
   virtual bool GetSize(int64* size_out) OVERRIDE;
